+++ src/whisper-utils/whisper-utils.h
@@ -11,8 +11,8 @@ void update_whsiper_model(struct transcription_filter_data *gf, obs_data_t *s);
 void shutdown_whisper_thread(struct transcription_filter_data *gf);
 void start_whisper_thread_with_path(struct transcription_filter_data *gf, const std::string &path);
 
-std::pair<int, int> findStartOfOverlap(const std::vector<whisper_token_data> &seq1,
-				       const std::vector<whisper_token_data> &seq2);
+std::pair<size_t, size_t> findStartOfOverlap(const std::vector<whisper_token_data> &seq1,
+					     const std::vector<whisper_token_data> &seq2);
 std::vector<whisper_token_data> reconstructSentence(const std::vector<whisper_token_data> &seq1,
 						    const std::vector<whisper_token_data> &seq2);
